- ScnPackage basics implemented. - Fixed type size problem in x86_64. - Exposed internal resource methods of CsCore (ScnPackage needed raw internals :( ) - Temporary skinned mesh support - loads as static meshes. - ScnModelInstance implemented.

+++ Engine/Source/Shared/Base/BcTypes.h
@@ -187,7 +187,7 @@ typedef	unsigned short				BcU16;
 typedef unsigned char				BcU8;
 
 typedef signed long					BcS64;
-typedef signed long int				BcS32;
+typedef signed int					BcS32;
 typedef	signed short				BcS16;
 typedef signed char					BcS8;
